# grading-platform-v2/docs/api/openapi-v2.yaml
openapi: 3.1.0
info:
  title: AI Grading Platform V2 API
  version: 0.2.0
  description: >
    V2-only API contract for extension-app and admin-console.
servers:
  - url: http://localhost:3000
    description: local
paths:
...
  /api/v2/settings:
    get:
      summary: List settings or query by key
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - in: query
          name: key
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Upsert setting
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingUpsertRequest"
      responses:
        "200":
          description: Updated
        "409":
          $ref: "#/components/responses/ApiError"
    delete:
      summary: Delete setting by key
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - in: query
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
        "409":
          $ref: "#/components/responses/ApiError"
...
  /api/v2/rubrics:
    get:
      summary: List rubrics or get by questionKey
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - in: query
          name: questionKey
          schema:
            type: string
        - in: query
          name: examId
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Upsert rubric
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RubricUpsertRequest"
      responses:
        "200":
          description: Updated
        "409":
          $ref: "#/components/responses/ApiError"
    delete:
      summary: Delete rubric by questionKey
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - in: query
          name: questionKey
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
        "409":
          $ref: "#/components/responses/ApiError"
  /api/v2/rubrics/generate:
    post:
      summary: Generate rubric draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RubricGenerateRequest"
      responses:
        "200":
          description: Generated
...
components:
  parameters:
    XActivationCode:
      in: header
      name: x-activation-code
      schema:
        type: string
    XDeviceId:
      in: header
      name: x-device-id
      schema:
        type: string
    IdempotencyKey:
      in: header
      name: idempotency-key
      schema:
        type: string
  responses:
    ApiError:
      description: Standard API error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
...
  schemas:
...
    SettingUpsertRequest:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value: {}
...
    RubricUpsertRequest:
      type: object
      required: [rubric]
      properties:
        questionKey:
          type: string
        rubric: {}
        examId:
          type: string
          nullable: true
        lifecycleStatus:
          type: string
          enum: [draft, published]
    RubricGenerateRequest:
      type: object
      properties:
        questionImage:
          type: string
        answerImage:
          type: string
        answerText:
          type: string
        questionId:
          type: string
        subject:
          type: string
        questionType:
          type: string
        strategyType:
          type: string
        totalScore:
          type: number
...
    ApiErrorEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
...